pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.2.0"
description = "difference calculator"
authors = ["Dmitry_Babokin <roodmannofficial@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
